+++ spec/models/user_spec.rb
@@ -42,6 +42,20 @@ describe User, "making up the URL name" do
   end
 end
 
+describe User, "banning the user" do
+
+  it 'does not change the URL name' do
+    user = FactoryGirl.create(:user, :name => 'nasty user 123')
+    user.update_attributes(:ban_text => 'You are banned')
+    expect(user.url_name).to eq('nasty_user_123')
+  end
+
+  it 'appends a message to the name' do
+    user = FactoryGirl.build(:user, :name => 'nasty user', :ban_text => 'banned')
+    expect(user.name).to eq('nasty user (Account suspended)')
+  end
+
+end
 
 describe User, "showing the name" do
   before do